Restructure project for Vercel deployment
