+++ vercel.json
@@ -1,0 +1,7 @@
+{
+  "functions": {
+    "api/*.py": {
+      "runtime": "@vercel/python@4.7.2"
+    }
+  }
+}
